Add hook to annotate more CSS classes on events
This allows us pass a prop `addEventClass` which can return an object or string (whatever `classnames()` accepts) with additional CSS classes to apply to the event.

This means we can for example colour-code events depending on some metadata. Better yet because it's done for each day, we can style different days of the same event differently if required.