Fix bug whereby events crossing year boundary could be hidden
Date maths is hard. :)

If you have an event that finishes in an earlier month in a later year, it could be hidden by the previous logic.

Feels like it should be possible to do this with moment().isBetween(), but it kind of gets a bit tricky because we don't check against a specific day.