Fix missing event data in handler callbacks
Fixes #24

There is currently a bug where event data will be missing when the (e.g. `onMouseOver`) callbacks trigger - I currently get `false` on occasion for the [`eventData` param in my callback](https://github.com/dptoot/react-event-calendar#properties).

I believe this is because `this.sharedArguments` is only computed and cached once, when the component is created, and never updated as the props change. So on each subsequent re-render, there is a chance the two will become out of sync.

This fixes it by updating `this.sharedArguments` each time the props update.

@calvinclaus